requirements.txt
numpy<2

vitaldb
dill
tqdm
statsmodels
python-dotenv
matplotlib
scipy
neurokit2
openpyxl
seaborn
scikit-optimize
scikit-learn
xgboost

fastparquet
pyarrow


neptune == 1.14.0
neptune-pytorch == 2.0.0

--find-links https://download.pytorch.org/whl/torch
torch == 2.0.1+cu118
--find-links https://download.pytorch.org/whl/torchaudio
torchaudio == 2.0.2+cu118
--find-links https://download.pytorch.org/whl/torchvision
torchvision == 0.15.2+cu118
